requirements.txt
torchvision==0.11.3
pytorch==1.10.2
tokenizers==0.10.3
numpy==1.21.5
tqdm==4.59.0
transformers==4.18.0
matplotlib==3.3.4
